I fixed the category and name saving to be saved as lowercase

+++ pages/api/categories/[id].js
@@ -1,5 +1,6 @@
 import connectDB from "../../../utils/connectDB";
 import Categories from "../../../models/categoriesModel";
+import Products from "../../../models/productModel";
 import auth from "../../../middlewares/auth";
 
 connectDB();
@@ -40,6 +41,12 @@ const deleteCategory = async (req, res) => {
     if (result.role !== "admin")
       return res.status(400).json({ err: "Authentication is not valid" });
     const { id } = req.query;
+    const products = await Products.findOne({ category: id });
+    if (products)
+      return res.status(400).json({
+        err: "Please Delete or edit all products with a relationship",
+      });
+
     await Categories.findByIdAndDelete(id);
     res.json({ msg: "Succes! deleted category" });
   } catch (err) {
